refactor: move unit tests into dedicated directory
