+++ tests/unit/test_generate_data.py
@@ -29,13 +29,13 @@ from instructlab.sdg.generate_data import (
 )
 
 # Local
-from .taxonomy import load_test_skills
+from ..taxonomy import load_test_skills
 
 TEST_SYS_PROMPT = "I am, Red Hat® Instruct Model based on Granite 7B, an AI language model developed by Red Hat and IBM Research, based on the Granite-7b-base language model. My primary function is to be a chat assistant."
 
 TEST_TAXONOMY_BASE = "main"
 
-TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "testdata")
+TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "..", "testdata")
 
 NUM_INSTRUCTIONS_TO_GENERATE = 10
