+++ tests/unit/test_registry.py
