+++ tox.ini
@@ -18,8 +18,8 @@ package = wheel
 wheel_build_env = pkg
 deps = -r requirements-dev.txt
 commands =
-    unit: {envpython} -m pytest {posargs:tests --ignore=tests/functional}
-    unitcov: {envpython} -W error::UserWarning -m pytest --cov=instructlab.sdg --cov-report term --cov-report=html:coverage-{env_name} --cov-report=xml:coverage-{env_name}.xml --html=durations/{env_name}.html -p no:unraisableexception {posargs:tests --ignore=tests/functional -m "not (examples or gpu)"}
+    unit: {envpython} -m pytest {posargs:tests tests/unit}
+    unitcov: {envpython} -W error::UserWarning -m pytest --cov=instructlab.sdg --cov-report term --cov-report=html:coverage-{env_name} --cov-report=xml:coverage-{env_name}.xml --html=durations/{env_name}.html -p no:unraisableexception {posargs:tests tests/unit -m "not (examples or gpu)"}
     functional: {envpython} -m pytest {posargs:tests/functional -m "not gpu"}
 
 # format, check, and linting targets don't build and install the project to
